Refactor project structure for forms and settings
