+++ Client/Extensions/ConfigurationExtensions.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using PoverkaWinForms.Settings;
 
 namespace PoverkaWinForms;
