+++ Client/Forms/Admin/SetPasswordForm.resx
