+++ Client/Forms/Verifier/MetersSetupForm.resx
