+++ Client/Services/TokenService.cs
@@ -2,7 +2,7 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Json;
 using System.Text.Json.Serialization;
-using PoverkaWinForms;
+using PoverkaWinForms.Settings;
 using PoverkaWinForms.Exceptions;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
