+++ Client/Services/UserService.cs
@@ -3,7 +3,7 @@ using System.Net.Http.Headers;
 using System.Net.Http.Json;
 using System.Collections.Generic;
 using System.Linq;
-using PoverkaWinForms;
+using PoverkaWinForms.Settings;
 
 namespace PoverkaWinForms.Services;
