+++ Client/Settings/RoleConstants.cs
@@ -1,6 +1,7 @@
-namespace PoverkaServer;
+namespace PoverkaWinForms.Settings;
 
 public static class Roles
 {
     public static readonly string[] All = { "Verifier", "Admin" };
 }
+
