+++ Server/Settings/RoleConstants.cs
@@ -1,7 +1,6 @@
-namespace PoverkaWinForms;
+namespace PoverkaServer.Settings;
 
 public static class Roles
 {
     public static readonly string[] All = { "Verifier", "Admin" };
 }
-
